+++ Sources/SuperwallKit/Paywall/View Controller/Web View/Message Handling/PaywallMessageHandler.swift
@@ -21,6 +21,7 @@ protocol PaywallMessageHandlerDelegate: AnyObject {
   func openDeepLink(_ url: URL)
   func presentSafariInApp(_ url: URL)
   func presentSafariExternal(_ url: URL)
+  func requestReview(type: ReviewType)
 }
 
 @MainActor
@@ -154,6 +155,8 @@ final class PaywallMessageHandler: WebEventDelegate {
       handleCustomEvent(name)
     case let .customPlacement(name: name, params: params):
       handleCustomPlacement(name: name, params: params)
+    case .requestStoreReview(let reviewType):
+      requestReview(type: reviewType)
     }
   }
 
@@ -367,6 +370,10 @@ final class PaywallMessageHandler: WebEventDelegate {
     )
     hapticFeedback()
     delegate?.openDeepLink(url)
+  }
+
+  private func requestReview(type: ReviewType) {
+    delegate?.requestReview(type: type)
   }
 
   private func restorePurchases() {